requirements.txt
streamlit
pandas
openpyxl
PyMuPDF
pytesseract
Pillow
discord.py
python-dotenv
scikit-learn
joblib